+++ VendorTracker/Models/Order.cs
@@ -6,12 +6,12 @@ namespace VendorTracker.Models
   {
     public string Title { get; set; }
     public string Description { get; set; }
-    public int Price { get; set; }
+    public float Price { get; set; }
     public string Date { get; set; }
     public int Id { get; }
     private static List<Order> _instances = new List<Order> { };
 
-    public Order(string title, string description, int price, string date)
+    public Order(string title, string description, float price, string date)
     {
       Title = title;
       Description = description;
